requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.13.0
asgiref==3.11.1
attrs==26.1.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.4
cbor2==5.9.0
celery==5.4.0
celery-types==0.26.0
certifi==2026.4.22
cffi==2.0.0
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.4.7
click==8.3.3
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cryptography==47.0.0
daphne==4.1.2
distro==1.9.0
Django==5.0
django-cors-headers==4.9.0
django-filter==25.1
django-stubs==6.0.3
django-stubs-ext==6.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djangorestframework-stubs==3.16.9
filelock==3.29.0
fsspec==2026.3.0
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
hyperlink==21.0.0
idna==3.13
Incremental==24.11.0
Jinja2==3.1.6
joblib==1.5.3
kombu==5.6.2
MarkupSafe==3.0.3
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
numpy==2.4.4
packaging==26.2
pgvector==0.3.2
pillow==12.2.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.12
py-ubjson==0.16.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
PyJWT==2.12.1
pyOpenSSL==26.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
PyYAML==6.0.3
redis==5.0.1
regex==2026.4.4
requests==2.33.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==3.0.1
service-identity==24.2.0
setuptools==81.0.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==4.57.6
Twisted==25.5.0
txaio==25.12.2
types-PyYAML==6.0.12.20260408
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
ujson==5.12.0
urllib3==2.6.3
vine==5.1.0
wcwidth==0.6.0
zope.interface==8.4
